added back the config for upload and download jobs
